Somewhat buff Cocytus, Vestibule, etc. water monsters
There's no point to spawning a big fish in Vestibule's Cocytus entry
or a big fish zombie in Cocytus itself, but it makes reasonable sense
to have some water monsters in these areas noticably defined by water.
As a result, this replaces the weakling water monsters and their zombies
with somewhat stronger monsters that still serve as fodder for their depth,
enough to be worth spawning but not enough to threaten, aside from lucky kraken.
The choices are also amphibious rather than water-bound, but kiting zombies
is again a non-issue for the depth.

Though there is almost no water in the other hells and almost all of said
water is flagged for no monster spawning, they too would get these zombies
if water spawns are permitted for any reason.

This also explicitly restrains the number of water monsters than spawn
within Vestibule's Cocytus entry, using a function to fake spawn weights.

The real solution would be custom zombie lists for hells to accompany
these random picks instead of just floors of dragon skeletons, due to some
underlying brokenness, but whatever.