Demonspawn blood saint: Legendary Destruction and support spells.
The support spells are Orb of Electrocution (as a monster spell),
Explosive Bolt (seen on the proposal as Bolt of Fireball), and Flash
Freeze.

The monster takes 10 damage to cast one of Orb of Electrocution, Crystal
Spear, Orb of Destruction, Ghostly Fireball, Explosive Bolt, or Flash
Freeze (not all equally weighted).

Some of the messaging could be improved, but the underlying framework is
present and accounted for.