Remove `drop_mode` setting and in-game toggle for `single` behavior
The setting itself went largely unused with a few exceptions for bot
configurations, and the toggle (either @ or Ctrl-D in the drop menu)
added nothing to the interface: It was intransparent and probably not
known to a single player out there. Like with the pickup menu, what was
old `multi` is the far superior interface, and unlike with pickup it
also has been the default for a long time.
For quickly (interface-wise) dropping items, there still exists `D` to
drop your last pickup, but even that is rather prone to input errors.