Adjust siren pulling behavior
Siren pulling is now irresistible if you are already mesmerized and
happens passively each turn (without her needing to renew her song),
but only triggers if the player did not themselves approach the siren
that turn. (The idea is that this prevents you sometimes jumping
multiple spaces in a single turn and makes the pulling behavior more consistent but also
more powerful).