A few more siren behavioral tweaks
Make sirens even more likely to prefer deep water than before (to
the point of being willing to move past the player to reach some, when
previously they never would if shallow was available closer to them).
Given how key it is to their offense, it seems important to prioritize
it further.

Also adjust siren pathfinding slightly so that they will attempt to
move around the player instead of stopping in front of them, if you
happen to bar their path.

Finally, remove siren's melee attack altogether. It almost never
works anyway and was harmless even when it did. It is better not to
give the illusion that she will actually hit something when she so
rarely does, I think.