Actually give monsters their intended ammo on creation
Because it checked the same code that governs whether a monster will
stop to pick something off the ground, monsters with ranged spells
would refuse to accept ammo given to them deliberately on creation.
Notably, this means that Harold never got any of the nets he was
supposed to have (and hasn't for years).

I don't really see why this code should be involved here, since
presumably if mon-gear tries to give a monster a specific item, it
does mean for the monster to have it, even if it would not otherwise
pick that item off the ground.