Demonspawn black suns: Black Mark.
Originally Black Hole in the proposal, but with a Singularity spell
in the works that's too misleading a name.

This spell empowers the caster and its allies to randomly drain speed,
skills, abilities, or magic on successful damaging melee attacks,
healing the ally making the attack.

This commit also gives black suns a (temporary?) spell set loosely based
in the proposal; they also get Malign Offering, which I see as fitting
both the necromancy theme and the ability of the caster to heal their
allies (through Black Mark).