Demonspawn warmonger: Sap Magic.
Based on Punishment of Excess from the proposal.

Casting spells while afflicted by this hex progressively hinders your
spellcasting success (or adds antimagic enchantment levels if you are a
monster).

Technically this could be extended to have an effect specific to
individual spell schools, but that would result in a lot of duplication
of durations and the like.