Spawn locales for the new enemies.
The demonspawn obviously spawn in Pan; base demonspawn have 1000 weight
each (same as 3s and 2s) and nonbase have 800 weight (80% of the base
weight, as with nonbase vs base draconians in Zot). They appear in bands
- 2-4 for base draconians, 4-7 for those led by nonbase (of which the
first follower has a 50% chance of being nonbase, the second has a 25%
chance of being non-base, and the remainder are base).

This also sneaks the worldbinder into Abyss, stealing 1/3 of the weight
of ynoxinuls per the original proposal.

The weights may need to be adjusted further.