Slightly revamp help header in Ctrl-F
The old one did really not help any player whatsoever, I tested with
several new ones. Nobody was able to find out what it tried to convey.

Hopefully this layout gets the important points across better now.
What should also be changed are some of the defaults themselves. This
might be addressed in a subsequent commit.