Add a simple monster implementation of manticore barb damage
I'm not sure this is really necessary and cannot operate the same as
the player's in any case, since monsters don't really know how to
stand still. Thus theirs will wear off even when stationary (though
moving accelerates the process as they take damage).

It might just be simpler to make the monster bleed or something, instead
of this extra code for an effect that will rarely happen or be meaningful
(though manticores have already been suggested as a possible player
summon several times, so....)