Remove decks of dungeons from Nemelex's gifting.
Not every deck has to be gifted by Nemelex, and while dungeons did have a
coherent flavour it did not have an obvious use, at least compared to the other
four. In particular, it was never well differentiated from wonders, and neither
were the items that increased its weight vis-a-vis the items that increased
wonders'.

Trowel, the one memorable thing about the deck, is added to wonders. Dungeons
is made more common as a random spawn to compensate for not being a gift.

Minor save bump.