Prevent some monsters from spawning with cursed weapons
I've been conservative with this functionality for the moment, and
applied it mainly to cases where switching between weapons is
critical to a monster's overall functioning. This includes satyrs
and naga snipers (who are archers that spawn wielding a melee weapon
and thus cannot even use their ranged focus at all if it happens to
be cursed), spriggan assassins (who rely on using their blowguns),
and Sonja (whose blowgun can I THINK start out cursed, which seems
a shame). There are maybe a couple other things that could benefit
from this, but it seems clear that archers shouldn't randomly be
deprived of their ranged weapons on generation.

The exact method of simply uncursing whatever item is generated,
after the fact, strikes me as somewhat hacky, but it is how vault
code to generate uncursed items currently works and the alternative
seems significantly more complex, so....