Throw Ds into random pan lord vaults, pan.des tweaks
Through functionizing up some basic chunks of to ostensibly also
serve as a clean-up in making a varied number of vault header styles
consistent and abbreviated, throw in some minor chances for demonspawn
to infiltrate the random pan lord vaults. The chances may be in fact
too low, but avoiding gimmick overload is just as important as
making sure each floor and the branch overall is unique with
these new branch-specific monsters. Further tweaking may be necessary.

(The stairs code management here is not playing greatly nicely
through inconsistency between four or three transit_pandemonium
a level in a way that seems to be seperate of how each vault
places a stair, so some further investigation may be warranted.
It should at least be much easier to address such a problem,
or the weights of inserted demonspawn or whatever.)

Also, slightly buff the weakest of the unrand pan lord vaults, and
throw in tiny chances for some guests into each unrand pan lord set.