Add a milestone at game start.
This lets Sequell, tourney scripts, CAO scoring pages, etc know about the
existence of every game that is in progress, not just the ones with a real
milestone.

It is currently not announced by any IRC bot except Ruffell, and RHF doesn't
seem to have enough games started per day for this to be a block (though
of course hopefully Ruffell will start muting them sometime soon).