Split the note_chat_messages option into note_chat_messages and note_dgl_messages.
Webtiles chat and DGL messages have different functionality and are used in
completely different ways. Webtiles chat doesn't necessarily involve the
player of the game being spectated (and indeed it often seems not to have
anything to do with the game), while DGL messages are invisible unless the
player chooses to read them and thus usually consist of a conversation
between the player typing in notes and a spectator. Not logging DGL
messages (which was recently made default) is thus usually undesirable
because it means logging only one side of a conversation (the player's).

The new option note_dgl_messages defaults to true.