Adjust OTR damage formula when cast by monsters
While effective and apparently balanced as a player spell, when used
against the player, OTR is underwhelming - in definite part due to
player's greater hp, ready access to !curing, and the fact that players
can kite with the spell while monsters will not. Overall it does little
direct damage while potentially poisoning too heavily (though all this
often amounts to is costing you a !curing after the battle is over).

With that in mind, I have rebalanced the effect against players in 3 ways:

1) It inflicts considerably more direct damage
2) It inflicts only half as much poisoning
3) The damage falloff by distance is greatly reduced for monster casters
   in general. (It made the effect too trivial if not engaged in melee with
   the caster, and monsters do not require any extra mechanisms to discourage
   engagement with the player)

Overall it should be be more threatening against players in an immediate
sense, while being less likely to leave you in the deep red afterward.