Buff corrupter's planerend summons
The pile of flat chances making an already random enough threat weaker
were getting on my nerves, and Pan should try to kill people
with a little less chance to drop in death yaks or whatever.
There are a lot of tweaks I'd like to make to the set still in the
wake of the Forest dispersal project, but for now this should *try*
to be a little more scary.