Make Ignite Poison properly useable by monsters, overhaul effect on players
This finishes some of the generalization made previously to this code
to make it less player-centric and adds tracer-like checks to allow
monsters to guess when it's worth casting (ie: will do more to their
enemies than their allies).

A lot of the player damage code left over from when it could hurt the
caster wasn't really suitable for a monster effect and so has been
removed or changed. It can no longer destroy items in your inventory
(including 100% of all poison ammo carried!) nor create flame clouds
based on this. Damage for higher levels of poisoning has been greatly
stepped down and poisoning no longer stacks with naturally poisonous
physiology (ie: being a naga or kobold) in terms of damage inflicted
on players.

I don't actually have a home in mind for this code directly, but I
plan to use it as the base for a single-targeted version of this
effect, and maybe it would be okay on player ghosts?