Ability menu always shows available abilities
This means no longer pressing `a?` or `a*` to see your abilities.
The list now is displayed right after pressing `a`.
There is no --more-- like for rarer prompts because the ability menu
title is a fine description. Another implication of omitting the
--more-- input is that certain muscle memory commands like `aa` still
work unchanged, and macros dealing with abilities also have an easier
time without it.

Had been the default for TOUCH_UI already. All code support for prompt
input is now removed.