A single-target variant of Ignite Poison for monsters
This is basically what it says on the tin: Ignite Poison delivered
to a single location via a (non-resistable) enchantment beam. The
damage inflicted is greater at the same power compared to normal
Ignite Poison, but making it non-smite-targeted is a considerable
limitation in the player's favor, of course.

Intended for Salamander Mystics.