Remove friendly fire from monster OTR
Even in a branch where most of the monsters naturally resist it, the
friendly fire can be problematic (around, say, anacondas, and now also
salamanders and shock serpents), and without being able to liberally
use this spell, naga ritualists suffer greatly. Let's just assume that
their explicitly esoteric magics have found an answer to this problem,
and go with an implementation that plays better.

This also simplifies tracer logic, since it can simply check only
for the presence of susceptible hostiles and not the balance between
those and friendlies.