Chaos champions: Rearrange the Pieces.
This is a close analogue of Xom's "rearrange the pieces" effect, only it
selects pairs of nearby creatures randomly to swap around rather than
swapping you with a nearby target first and then swapping nearby
monsters indiscriminately; every nearby creature here is guaranteed to
be repositioned (unless there are an odd number of creatures in view, in
which case one will be left out). There are also no confusion effects.

Chain of Chaos is now unused; it's been suggested it be used for a Xom
effect, but I'll leave that to others to figure out.