Add Wizard mode commands to identify/unidentify item types and level items.
This binds the commands to & y/Y.  We have & i/I, but that only affects
inventory items.  For various testing purposes, it's nice to be able to
mass-identify/unidentify all item types.  This command also
identifies/unidentifies all inventory items on the level, since items can refuse
to unidentify if they exist on the floor, and because level-item identification
is also a convenient thing to have.  We don't change the identification status
of off-level items.