Reduce extreme (buggy) steam spread rate from fire clouds over water
Not only are the absolutely massive explosions of steam created by flame
clouds over water somewhat annoying to have to sit through whenever you
(maybe inadvertently) cause them, but it looks like they were probably
unintentional in the first place. The steam cloud's spread rate (in just
this one context) is left unfilled at a default of -1. Of course, spread_rate
is an unsigned 8-bit integer.... (it should be noted that steam's NORMAL
spread rate is 22)

If you use wizmode, you can see that even a single flame cloud can produce
steam at up to 15 or so spaces away from the original source, which is rather
insane. I have reduced the spread rate to normal (and increased the generation
rate just a touch), which I think produces far less intrusive effects.

I'm not fully sure of the numbers, which might need further adjusting, but
combined with the last commit, I think it's a definite quality-of-life
improvement in Swamp for both people being attacked by vapours and those using
fire and lightning themselves.