Slightly reduce wisp band size, rarely spawn vapours with them
Pairing vapours with wisps is an idea that was tried a while back,
and then canned due to the forest fires vapours were prone to starting.
With that now made less of a problem (and vapours overall improved from
their old state back then, too), I think it's worth trying again.