Make vine stalker enemies less of a dummy.
They get a normal attack in addition to the bite; if they're used as
spellcasting enemies, they can remove the effects of antimagic through
their antimagic bite just as the player version can. This will probably
be more relevant if/when there's a version of the enemy that can cast
spells

Actually, this will be more relevant in general if/when they actually
spawn somewhere (Swamp and/or Snake have been suggested for good
candidates for this, probably moreso Swamp).