Adjust Shoals spawn table
A similar edit as with Swamp: monsters from the forest dispersal
are introduced and the weights of some others adjusted. Bats are
far less common, hippogiffs are replaced with griffons increasingly
by depth, high tier merfolk are much more common outside of Shoals:5
itself. Water nymphs are added as a mid-tier spawn, with rarer fauns
as low ones and then even rarer satyrs deeper. Water elementals can
spawn randomly to break up some of the monotony of filler spawns,
along with an ocassional wind drake.

For a few reasons I am somewhat less pleased with this than the
Swamps weights, potentially because Shoals had a larger band lacking
in many meaningful encounters and with fewer things to fill it with.
I expect there will be more changes to come in future.