Depths population table adjustments
Much less sweeping than Swamp and Forest, understandably. This adds
spriggan air mages and berserkers as mid-tier encounters, and defenders
as rarer high-tier ones, as well as making a few other adjustments
here and there. Deep trolls are rarer, flayed ghosts are somewhat more
common, and modest changes to a few other things I felt were a bit too
common or rarer.

(Defenders were added to the list twice with different ranges and
weights. I am not sure how better to reconcile the fact that all
weights in the OOD range are so incredibly low (and defenders make
sensible solo vault spawns from this list) with wanting them to
sometimes spawn OUTSIDE of vaults, as their band is a point of interest
for them. Possibly it would make more sense if the scale for depths
greater than 6 were not so different than depths 6 or below, but
adjusting this would be a lot more work right now)