Stack nonbase demonspawn HD sizes, AC, EV on base demonspawn values.
Previously the size of hit dice used for nonbase demonspawn was taken
directly from their base type; the values given for those specific types
were now meaningless.

This changes the previously-unused numbers to be bonuses on top of the
base demonspawn type, so that, say, warmongers get their intended HP
bonus on top of other nonbase types (cf. e7a4d24).

This is also taken into account for the average HP calculation used for
ordering monsters in the monsters list.