Let siren song actually trigger
Somehow the code block that would cause the drowned soul call to
begin mysteriously vanished between the effect originally being coded
and tested and now. Even I am not sure what happened to it, and the
reflog hasn't helped. My best guess is that somehow it was accidentally
deleted along with nearby debugging code, maybe?

I think the original code determining song duration here was more
involved, but I cannot remember it. This will at least allow it to
function, and might well be good enough as-is.