Buff demonspawn enemy melee damage
Even if they're meant to be 'simple' fodder, the base demonspawn
generally had weaker melee than some (melee-focused) 4s, which is
silly when they're supposed to dillute Pan's slice of 3s and come
in bands with monsters that explicitly want good melee accompanying;
this gives just a minor buff to be able to approximate those melee 4s.
Also buffing the melee damage of the classes due to some non-melee
2s dangerous by spells still having better damage; most notably,
this should help black suns and chaos champions have a bit more
raw threat to themselves and further up warmongers to approximate
something of stone giant damage or whatever. Extended can take it.