New Zig Set: negative energy!
It's as much an element as fire, ice, and air, and this gets to
spam some current side-attractions for some interesting gains:
slowing-melee mp-draining-melee profane servitors and some
wonderful shadow fiend numbers. Use your damn rN+ already.

(If you're undead, this still isn't a breeze:
 three monsters serious monsters here have dispel undead.
This could really use a tiles icon for black mark, but whatever,
it's good enough to just always assume that servitor has it.)