Functionality for branch zombie lists (currently unused).
This is designed with the Hell zombie problem in mind, though recent
tweaks have also caused, say, Crypt to have problems with generating
zombies outside of _place_assorted_zombies() - right now it only
naturally generates deep dwarf zombies and elf zombies.

If you don't want to use it for a branch, just set the zombie list to
the basic monster list and the current behaviour is retained. That's the
case for all branches currently.