Unique: Vashnia, leader of the elite marksnaga squad.
An HD 16 naga sharpshooter who also comes with various Blink spells,
inherent Phase Shift, and a Shroud of Golubria that she can renew if it
gets broken.

Appears on Snake:2-.

Historical note: Vashnia's original design had Disjunction as an escape
spell; currently the escape spell is Blink Other, which accomplishes
basically the same thing but with less visual impressiveness. Perhaps at
a later date the possibility of this can be revisited.