Make a custom generic zombie set for Crypt vaults, extended
The previous method of zombie creation, based on in-branch spawns
as are valid for zombies followed by whatever matches the same
absdepth, is inherently not going to work well when it means most
hells are stuffed with dragon zombies and Gloorx's level has only
demonspawn skeletons. This set-up will serve as the general zombie
base set for such spawns in Hells, Pan, Abyss, Tomb, and vault-placed
zombies in Crypt.

The weighting of monsters as bases is 26.25% giants, 28.75% dragons,
37.5% Lair and Lair roulette representatives (with three bases
alotted per branch), and 7.5% mostly-harmless player species
aside from the four with zero non-vault representation or the
six that aren't living (for flavour purposes). This should be
a reasonable enough and understandable enough set-up for what
one can find as a zombie in most of extended, and represents
most of the workable zombies.

(Cocytus thus has some weird settings where some zombies will
be patrolling water and some won't, but it'd be quite difficult
to ever notice this in-game considering it's Cocytus and thus
a Hell to be dived quickly through.)