Split M_SHADOW from M_GLOWS_LIGHT.
Historically, M_GLOWS_LIGHT was originally a hardcoded list of monsters
that can't be backlit: the list was mostly glowing creatures, but also
some inherently shadowy ones.  When the use of the flag was subsequently
expanded, it be

Now shadowy creatures:

 * Can go invisible (even more shadowy)
 * Have a stealth bonus, not malus.
 * Are described as "wreathed in shadows" rather than "outlined in
   light".

Give the flag to shadow fiend, shadow dragon, phantasmal warrior,
spectral weapon, and shadow fiend.  Don't give it to spectral things,
which are explicitly described as glowing.