Buff the two plain no-Lernaean Swamp:$ vaults with new Swamp jerks
It is somewhat sad that the least gimmicked, no-outsiders
Swamp:$ rune vaults are the weakest of the lot by far considering
said context of most of Swamp:$s relying on monsters that don't
even spawn in Swamp. As such, with newly-buffed spawns throughout
the branch, grab the three of the most dangerous ones (thorn hunter,
spriggan druid, and treant) and put in two of any combination of
these for "swamp" and swamp_alternative alongside the default
mess of hydras and swamp dragons. Each is not only a mean threat
by itself but also synergize well with company, so it should
play fine.