Fix compilation, five vaults (Cheibrodos, #8009).
An overflow altar for Chei and another for Ash,
a minivault with a gnoll and a minivault with an orc knight,
a 098 ORIENT: float vault. Minor edits from the submission
for most, but the floating vault gets a big NSUBST to prevent
it from placing something like 70 monsters for one D:14 vault.

Also, fix placing 'treants' instead of 'shambling mangroves', whoops.