Automatically identify chopped hides.
Add an extra step in _create_monster_hide to automatically identify
chopped dragon and troll hides. Hides are already always generated as +0
uncursed, so identifying them automatically removes both a spoiler and an
unnecessary step of forcing the player to wear or ?identify dragon
armours that they have created specifically to enter a trove.