Buff Asterion's HP.
Up to 150 from 120.

It seemed really odd that a unique born and bred for killing and
inherently more robust than humans only had as much HP as similarly
fighty humans with about the same HD (cf. Maud, HD 13 and 118 HP, or
Frances, HD 14 and 121 HP); 120 HP is also at the top end of the range
for fire giants or frost giants (both HD 16).

This is partially motivated by initial reports that Asterion isn't very
interesting to fight (yet?). If this proves to be wrong, and/or this
goes too far in the other direction, we can always tinker with him
further.