Halve base Dith piety gain from kills.
i.e. the additive modifier for fiery/illuminating enemies is still the
same, but the base gain for kills is cut in half.

Prior to this it's been too easy to earn and maintain Dith piety, which
is bad in light of the power of some of Dith's abilities.

This may be *too* much of a nerf - we'll have to see how this pans out.