Make failure_rate_to_int return more sane values.
The only effect of this is that the "failure rate" shown when examining a
100% fail spell should be a bit more useful now, since it is now linear
in stuff like intelligence and skills rather than cubic. Also, the cubic
formula overflowed an int and gave garbage like "Guaranteed Fail: 1%"
in extreme cases.

I still don't much like displaying >100% failure rates in the first
place, though. It's intrinsically nonsensical and of very limited use to
the player.