Make "every 20 turns" effects happen less predictably.
This introduces event timers that schedule the next time the code
handling one of these effects - including such things as hell effects,
glowsplosions, and divine retribution - to happen a random amount of
time after the current time the event runs.

Most things are called every 100-300 aut (or a multiplier thereof); a
minimum delay between instances seems appropriate, as does a maximum.

I'm not completely satisfied with this implementation, but it gets the
job done.