Revamp rod of destruction.
The rod of destruction {lightning bolt, fireball, iron shot} is pretty boring,
especially compared to the number of rods with unique spells. The new rod of
of destruction has only one (L5) spell, "Random Bolt," which zaps a random
bolt from the list of venom, draining, fire, cold, quicksilver, lightning,
or crystal.

The rod will prompt for self-harm if it could bounce, even if the bolt that
comes out won't bounce (e.g. fire on rock). This isn't optimal but there's
no really other way to do it, as far as I can tell.