Block plain/base demonspawn from polymorph/shapeshifters.
The unholy mix of demons and mortals is already a holiness swap
that polymorph cannot cross, and their weird hybrid state as well
as the whole current MH_LIVING status does not make it plausible
that shapeshifters could just instantly copy them; the status
is more mechanical than functional for this particular bit.

Blocking not-actually-used player-species-dummy-monsters from
vault use or polymorph is a worthwhile idea, but it's muddled
a bit by this most recent incursion of Ds onto shapeshifters;
this should make the overall change easier to go through,
especially with confusion over fasceted-demonspawn as monsters
or not compared to the non-fasceted, dummy "demonspawn" enemy.