Don't crash when sirens sing near occupied water.
If there is deep water in LOS radius of the siren, but all the cells
have an actor, siren_song would crash as it tried to access element 0 of
the empty deep_water array.  Likewise if there were unoccupied cells but
we tried to place more drowned souls than there were cells.

See for example:
  http://dobrazupa.org/morgue/johnnyzero/crash-johnnyzero-20140208-225441.txt