Don't penalize monster's spectral weapons with half hd.
Currently monster::skill() reports weapon skill as half of hd for
non-fighter monsters.  This is only used for spectral weapon, where it
results in specters with half-monster-hd and hence poor to-hit.  This
changes monster::skill() to not penalize based on M_FIGHTER.  This
doesn't affect damage or other aspects of spectral weapons, which are
based on the weapon properties.  A boost for M_FIGHTER monsters using
spectral weapon (or otherwise somehow using monster::skill()) can be
given in the future if need arises.

This commit also cleans up a comment.