Remove arbitrary gear slot connections.
Specifically, body armour does not require taking on/off cloaks, and putting on
and removing rings with cursed gloves is allowed.

This was a constant source of nonsense: why can you not put on rings with
cursed gloves, but you can put on gloves with cursed two-handers? Why do cloaks
block armour but gloves don't block full plate? The only possible motivation
was "realism," but that's a weak argument in general and having the slots not
being connected is not particularly unrealistic.

[MarvinPA: The one place this was occasionally relevant was under Ashenzari,
but even there the result was just that you would always prefer to curse
boots/helmets/body armour over cloaks/gloves to avoid the extra
micromanagement.]