Don't leak information on out-of-sight squares in xv (#8109)
This allowed you to identify altars and portals detected by magic
mapping, even though map_knowledge explicitly had DNGN_UNKNOWN_ALTAR.
Even worse, xv on an unseen (even unmapped) region of the level would
tell you exactly what feature was that that spot.

This used to be prevented by xv only working in LOS.  Now we check
whether the real feature is the same as the remembered one, and if
not use the base description for the remembered feature.  If they
were the same, we look up the raw_feature_description for the
location, so that vault feature renames are applied.

The solution isn't perfect: it still can still leak whether a remembered
square has changed, detectable by the lack of feature renames (so only
in a few rare locations).  Probably map_knowledge needs to store the
rename as well to fix that issue.