Set most 'excl_rad = -1' to 0
I'm not convinced that this is full necessary in terms of function,
considering how most of these vaults are in places questionable
to autoexplore and most of these generators would only avoid a prompt
into a cloud if the cloud was directly disturbed by e.g. ice/fire
or fan of gales, but whatever. Cloud machines are already awkward
in a lot of ways, with being mostly tied to item destruction instead
of damage past Lair, not providing something to directly confront
in an inconsistent fashion against Crawl's main focus of fighting
monsters, not an easily automated process, etc; I'm hardly about
to defend these vaults using them for flavour purposes.