Mercenary shops.
Like the Mercenary card, pay an amount of gold roughly determined by the
merc's XP value to get a permanent ally.

The monster list is taken from that of the Mercenary card with a few
additions.

The implementation of this is via a dummy OBJ_MERCENARY item type. I
don't think any huge issues will result from this, but be warned.

Some currently wonky behaviour:
* For MAKE_GOOD_ITEM (which is true for one in four shop items at
  present), you end up with a top tier mercenary. Depending on where
  these shops appear this may or may not be desirable.
* If there isn't room to place the merc, your money is wasted. Should
  there be a refund or similar?