Set timers properly when transferring old saves.
We were using you.elapsed_time before unmarshalling it, leading to
triggers scheduled for 0.  As a result, in a sufficiently long
transferred game, monsters would no longer leave corpses (rather, they
would rot away immediately).

Besides filling in missing timers with the correct time, also fix up any
next-trigger-in-past timers that resulted from this problem.  If the
save did not come from a version with the buggy unmarshalling, assert
instead (as that indicates an unknown problem that should not be papered
over).