Adjust shadow mimic HD / damage calculations.
The base attack damage is 0 unless you're unarmed, in which case it's
the greater of 1 or half your unarmed combat skill; that is, all of the
damage comes from the weapon type.

Also, its melee/ranged HD is now half your skill value in the relevant
weapon. (This may be too low.)