Revert global spriggan monster nerf
Seperating monster spriggan move/action speeds ostensibly for parity
with player spriggans is also a massive damage nerf to an entire class
of monsters, many of which do not seem to warrant one. And if some do,
that is better handled on a case-by-case basis, addressing specific gameplay
issues. This change was both overly broad and harsh.

This reverts commits 6f0844da8a28793ebe3b81f9762de0869deccb96
and c004371e41787b2844a6433cd44f2b96e22d98c0.