Somewhat unnerf deep elf summoners
These were hit especially hard by both the summon cap and the
abjure-on-death change, since they are squishy and rely on their
summons for all of their threat (and their summons were of a
somewhat swarming style that was hard to achieve with a cap of
only 3)

For comparison, they were about on parity with deep elf conjurers
in 0.13, yet score less than 1/4 of their kills in 0.14 so far (and
less than even fighters, who are the weakest of all deep elves,
supposedly).

This raises the cap to 5, adds an extra slot of summon vermin,
gives them some more hp, and very slightly adjusts summon weightings.
I expect they will still be much weaker than in 0.13 anyway, but
this might help a little.